Fix: Type conversion error and improve FlareSolverr reliability
COMPILATION FIX:
- Fixed int to int64 type conversion error in channel_file.go:622
- Added explicit int64() cast for ch.Filesize when file is already deleted

FLARESOLVERR IMPROVEMENTS (based on deep research):
- Increased startup timeout to 180s (FlareSolverr can take 30-60s to initialize)
- Added BROWSER_TIMEOUT=40000 environment variable (40s per request)
- Added HEADLESS=true for better performance in CI
- Improved health check with HTTP status code verification
- Added exponential backoff retry strategy
- Better container status monitoring
- Added 5s stabilization delay after API becomes ready
- Enhanced error reporting with container status and logs

DOCKER COMPOSE:
- Added proper healthcheck configuration
- Set start_period to 60s (allows time for initial startup)
- Used condition: service_healthy for proper dependency management
- Ensures recorder only starts after FlareSolverr is fully ready

Based on research from:
- FlareSolverr official documentation
- GitHub Actions service container best practices
- Docker Compose health check patterns

+++ channel/channel_file.go
@@ -619,7 +619,7 @@ func (ch *Channel) logUploadToDatabase(filePath, gofileLink string) error {
 	} else {
 		// File might have been deleted already, try to get size from channel state
 		ch.fileMu.RLock()
-		fileSize = ch.Filesize
+		fileSize = int64(ch.Filesize)
 		ch.fileMu.RUnlock()
 	}
